Renamed back Coercion.ts with correct capitalized C
